chore: move storage methods to storageFunctions class
