+++ src/main/java/com/limechain/runtime/hostapi/functions/StorageHostFunctions.java
@@ -5,25 +5,26 @@ import lombok.experimental.UtilityClass;
 import org.wasmer.ImportObject;
 import org.wasmer.Type;
 
+import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.List;
 
 @UtilityClass
-public class StorageHostFunctions {
+public class StorageHostFunctions extends HostApi {
 
     public static List<ImportObject> getFunctions(){
         return Arrays.asList(
                 HostFunctions.getImportObject( "ext_storage_set_version_1", argv -> {
-                    HostApi.extStorageSetVersion1((Long) argv.get(0), (Long) argv.get(1));
+                    extStorageSetVersion1((Long) argv.get(0), (Long) argv.get(1));
                 }, List.of(Type.I64, Type.I64)),
                 HostFunctions.getImportObject( "ext_storage_get_version_1", argv -> {
-                    return List.of(HostApi.extStorageGetVersion1((Long) argv.get(0)));
+                    return List.of(extStorageGetVersion1((Long) argv.get(0)));
                 }, List.of(Type.I64), Type.I64),
                 HostFunctions.getImportObject( "ext_storage_read_version_1", argv -> {
                     return argv;
                 }, List.of(Type.I64, Type.I64, Type.I32), Type.I64),
                 HostFunctions.getImportObject( "ext_storage_clear_version_1", argv -> {
-                    HostApi.extStorageClearVersion1((Long) argv.get(0));
+                    extStorageClearVersion1((Long) argv.get(0));
                 }, List.of(Type.I64)),
                 HostFunctions.getImportObject( "ext_storage_exists_version_1", argv -> {
                     return argv;
@@ -31,7 +32,7 @@ public class StorageHostFunctions {
                 HostFunctions.getImportObject( "ext_storage_clear_prefix_version_1", argv -> {
                 }, List.of(Type.I64)),
                 HostFunctions.getImportObject( "ext_storage_clear_prefix_version_2", argv -> {
-                    return List.of(HostApi.extStorageClearPrefixVersion2((Long) argv.get(0),(Long) argv.get(1)));
+                    return List.of(extStorageClearPrefixVersion2((Long) argv.get(0),(Long) argv.get(1)));
                 }, List.of(Type.I64, Type.I64), Type.I64),
                 HostFunctions.getImportObject( "ext_storage_append_version_1", argv -> {
                 }, List.of(Type.I64, Type.I64)),
@@ -56,5 +57,38 @@ public class StorageHostFunctions {
                 HostFunctions.getImportObject( "ext_storage_commit_transaction_version_1", argv -> {
 
                 }, HostFunctions.EMPTY_LIST_OF_TYPES));
+    }
+
+
+    public static void extStorageSetVersion1(long keyPtr, long valuePtr) {
+        byte[] key = getDataFromMemory(keyPtr);
+        byte[] value = getDataFromMemory(valuePtr);
+
+        repository.save(new String(key), value);
+    }
+
+    public static int extStorageGetVersion1(long keyPtr) {
+        byte[] key = getDataFromMemory(keyPtr);
+        Object data = repository.find(new String(key)).orElse(null);
+        if (data instanceof byte[] dataArray) {
+            return putDataToMemory(dataArray);
+        }
+        return 0;
+    }
+
+
+    public static void extStorageClearVersion1(long keyPtr) {
+        byte[] key = getDataFromMemory(keyPtr);
+        repository.delete(new String(key));
+    }
+
+    public static long extStorageClearPrefixVersion2(long prefixPtr, long limitPtr) {
+        String prefix = new String(getDataFromMemory(prefixPtr));
+        int limit = new BigInteger(getDataFromMemory(limitPtr)).intValue();
+
+        int deleted = repository.deletePrefix(prefix, limit);
+        //TODO: Count how many are left?
+
+        return 0;
     }
 }
